+++ src/main/java/com/team1/backendApi/repository/UserHistoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.team1.backendApi.model.UserHistory;
 
 @Repository
-public interface UserHistoryRepository extends JpaRepository<UserHistory, Integer> {
+public interface UserHistoryRepository extends JpaRepository<UserHistory, Long> {
     
 }
